+++ R/makevar_functions.R
@@ -1,19 +1,21 @@
-#' Use variable metadata to find variables containing a given string.
+#' Use variable metadata to find variables containing a given string
 #'
-#' Return a list of variable names or data frame of variable metadata containing a given string.
-#' This can be used create a list of variables to later pass to getCensus, or a data frame
-#' documenting variables used in a given project.
+#' Return a list of variable names or data frame of variable metadata containing
+#' a given string. This can be used create a list of variables to later pass to
+#' getCensus, or a data frame documenting variables used in a given project.
 #'
-#' @param name API name - e.g. acs5. See list at https://api.census.gov/data.html
-#' @param vintage Year of dataset, e.g. 2014 - not required for timeseries APIs
-#' @param find A string to find in the variable metadata
-#' @param varsearch Optional argument specifying which fields to search. Default is "all".
-#' Options are "all", "name", "label", or "concept".
-#' @param output Optional argument, specifying output to "list" or "dataframe". Default is "list".
+#' @param name API programmatic name - e.g. acs/acs5. Use `listCensusApis()` to
+#'   see valid dataset names.
+#' @param vintage Vintage (year) of dataset. Not required for timeseries APIs.
+#' @param find A string to find in the variable metadata.
+#' @param varsearch Optional argument specifying which fields to search. Default
+#'   is "all". Options are "all", "name", "label", or "concept".
+#' @param output Optional argument, specifying output to "list" or "dataframe".
+#'   Default is "list".
 #' @export
 #' @examples
-#' # Return a list, and then use getCensus function to retrieve those variables
 #' \dontrun{
+#' # Return a list, and then use getCensus function to retrieve those variables
 #' myvars <- makeVarlist(name = "timeseries/poverty/saipe",
 #'   find = "Ages 0-4",
 #'   varsearch = "label")